+++ lib/hcloud/entities/deprecation.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class Deprecation < Entity
+    attribute :announced, :datetime
+    attribute :unavailable_after, :datetime
+  end
+end