+++ lib/hcloud/resource_type.rb
@@ -90,6 +90,7 @@ ActiveModel::Type.register(:certificate, HCloud::ResourceType.Type("HCloud::Cert
 ActiveModel::Type.register(:certificate_status, HCloud::ResourceType.Type("HCloud::CertificateStatus"))
 ActiveModel::Type.register(:datacenter, HCloud::ResourceType.Type("HCloud::Datacenter"))
 ActiveModel::Type.register(:datacenter_server_type, HCloud::ResourceType.Type("HCloud::DatacenterServerType"))
+ActiveModel::Type.register(:deprecation, HCloud::ResourceType.Type("HCloud::Deprecation"))
 ActiveModel::Type.register(:dns_pointer, HCloud::ResourceType.Type("HCloud::DNSPointer"))
 ActiveModel::Type.register(:error, HCloud::ResourceType.Type("HCloud::Error"))
 ActiveModel::Type.register(:firewall, HCloud::ResourceType.Type("HCloud::Firewall"))